requirements.txt
Flask==2.3.2
python-gitlab==3.14.0
PyYAML==6.0
gunicorn==20.1.0
python-dotenv==1.0.0